requirements.txt
absl-py==0.2.2
astor==0.6.2
bleach==1.5.0
certifi==2018.4.16
coverage==4.5.1
dj-database-url==0.5.0
Django==2.0.6
django-environ==0.4.4
django-heroku==0.3.1
gast==0.2.0
grpcio==1.12.1
gunicorn==19.9.0
h5py==2.8.0
html5lib==0.9999999
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
Markdown==2.6.11
numpy==1.14.5
pandas==0.23.1
protobuf==3.6.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
whitenoise==3.3.1
dj-database-url==0.4.0
Django==2.0.7
gunicorn==19.5.0
psycopg2==2.7.1
whitenoise==4.0b4
django_heroku
django-filetransfers
